apply RAM optimizations (generated using Claude)
**`streamv2v/inference_common.py`**
- `load_generator_state_dict`: replaced `torch.load(..., map_location="cpu")`
with a version that tries `mmap=True, weights_only=True` first
(falls back to the original if PyTorch is too old)
- Same function: replaced the dict comprehension that added `"model."`
prefixes with an in-place `.pop()` rename to avoid a second full copy of all
tensors
- Same function: added `checkpoint.pop(key)` + `del checkpoint` to drop
optimizer state and other top-level keys before processing generator weights

**`streamv2v/inference.py`**
- `load_model`: added `del state_dict; gc.collect()` after `load_state_dict`
completes, so the CPU weight copy is freed before the GPU move happens

**`models/wan/wan_wrapper.py`**
- `WanTextEncoderWrapper.__init__`: changed `dtype=torch.float32` to
`dtype=torch.bfloat16` on the UMT5-XXL text encoder — this is what was
causing an OOM on low VRAM cards, as float32 puts the encoder at ~19GB

**`models/wan/causal_stream_inference.py`**
- `prepare()`: added `self.text_encoder.to('cpu')` +
`torch.cuda.empty_cache()` immediately after the text encoder runs, so it
gets offloaded back to CPU once the prompt embeddings are produced and
doesn't sit in VRAM for the rest of the session

+++ models/wan/wan_wrapper.py
@@ -43,7 +43,7 @@ class WanTextEncoder(TextEncoderInterface):
         self.text_encoder = umt5_xxl(
             encoder_only=True,
             return_tokenizer=False,
-            dtype=torch.float32,
+            dtype=torch.bfloat16,  # was float32 (~19GB) — bf16 halves this to ~9.4GB
             device=torch.device('cpu')
         ).eval().requires_grad_(False)
         self.text_encoder.load_state_dict(
